requirements.txt
discord.py[voice]
wavelink
easy_pil